requirements.txt
aiohttp==3.6.3
aiohttp-jinja2==1.0.0
aniso8601==8.0.0
appdirs==1.4.4
async-timeout==3.0.1
discord.py==1.5.1
PyNaCl==1.3.0